requirements.txt
boto3==1.9.153
botocore==1.12.153
Django==2.1
django-debug-toolbar==1.11
django-extensions==2.1.7
django-storages==1.7.1
docutils==0.14
jmespath==0.9.4
Pillow==6.0.0
psycopg2-binary==2.8.2
pygraphviz==1.5
python-dateutil==2.8.0
pytz==2019.1
s3transfer==0.2.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.3
